Refactor: Move SpatialIndex to farm.core.spatial
