+++ farm/core/spatial_index.py
@@ -1,9 +1,21 @@
-"""Compatibility shim for spatial indexing.
+"""Compatibility shim for spatial indexing (DEPRECATED).
 
-This module remains for backward compatibility. The implementation has been
-moved to the `farm.core.spatial` package. Importing from here will continue to
-work, but new code should prefer `from farm.core.spatial import SpatialIndex`.
+Deprecated: This module remains for backward compatibility only.
+The implementation has moved to `farm.core.spatial`.
+
+Please update imports to:
+    from farm.core.spatial import SpatialIndex, Quadtree, SpatialHashGrid
+
+This shim will be removed in a future release.
 """
+
+import warnings
+
+warnings.warn(
+    "farm.core.spatial_index is deprecated; use farm.core.spatial instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 from .spatial import (
     SpatialIndex,
